[cleanup] moved type file into types directory. Added readme for types and components directory
