+++ src/components/Readme.md
@@ -1,0 +1,8 @@
+# Componenets directory
+
+Please use this directory for componenets.
+Some design consideration for componenets:
+
+1. Keep the components are simple as possible
+2. Components should generally manage its own state
+3. Initail states to be received by props
