+++ src/components/types/Readme.md
@@ -1,0 +1,4 @@
+# Types directory
+
+Where possible, types should be grouped together in the same file.
+As a convension, type names should be capitalised while the filename should be camel cased.
